requirements.txt
pandas
streamlit
plotly-express==0.4.1